feat: Implement Hybrid Search basic planner (#547)

+++ rs/index/src/query/iter.rs
@@ -1,7 +1,7 @@
 pub trait InvertedIndexIter {
-    fn next(&mut self) -> Option<u64>;
+    fn next(&mut self) -> Option<u128>;
 
-    fn skip_to(&mut self, doc_id: u64);
+    fn skip_to(&mut self, doc_id: u128);
 
-    fn doc_id(&self) -> Option<u64>;
+    fn doc_id(&self) -> Option<u128>;
 }
